+++ test/sources/asm-0.7.5.sol
@@ -1,0 +1,9 @@
+pragma solidity >=0.7.5;
+
+contract Asm {
+    function foo(bytes calldata b) external {
+        assembly {
+            let z := b.length
+        }
+    }
+}