+++ test/sources/asm-0.7.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.7;
 contract Asm {
     uint s;
 
-    function foo() external {
+    function foo(bytes calldata b) external {
         assembly {
             let x := s.slot
             let y := s.offset