+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.4.58 (2024-08-26)
 
 - Fixed `YulFunctionDefinition.parameters` and `returnVariables`: made optional, used when empty.
+- Fixed `InlineAssembly.externalReferences[].suffix`: added `'length'` option.
 
 ### 0.4.57 (2024-07-16)